+++ src/render/integrator/mod.rs
@@ -20,7 +20,7 @@ pub trait Integrator {
         scene: &Scene,
         intersection: &SceneIntersection,
         sampler: &mut dyn Sampler,
-        depth: usize,
+        depth: i32,
     ) -> Spectrum;
 
     fn specular_reflection(
@@ -28,7 +28,7 @@ pub trait Integrator {
         scene: &Scene,
         intersection: &SceneIntersection,
         sampler: &mut dyn Sampler,
-        depth: usize,
+        depth: i32,
     ) -> Spectrum {
         let outgoing = -intersection.info.ray.direction;
 
@@ -67,7 +67,7 @@ pub trait Integrator {
         scene: &Scene,
         intersection: &SceneIntersection,
         sampler: &mut dyn Sampler,
-        depth: usize,
+        depth: i32,
     ) -> Spectrum {
         let outgoing = -intersection.info.ray.direction;
